pyproject.toml
[project]
name = "brain-hermes"
version = "0.1.0"
description = "Hermes MVP for Syncthing-backed shared memory workflows"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.122.0",
  "pyyaml>=6.0.3",
  "uvicorn>=0.38.0",
  "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
